Allow a custom suffix for versions
Adds the ability to specify a custom --suffix that gets tacked on
to the version number pulled from package.json.

This is useful when, for example, you cut a release branch and
begin QA testing on that release branch while devs continue merging
to master.

In this scenario the release branch is frozen in time while testing
proceeds. Sometimes (hah!) we need to pull a hotfix in from the
master branch. Now if we pull in the hotfix and the version gets
automatically bumped, we have two duplicate version numbers with
(potentially) very different source code. This is confusing to just
about everyone.

Providing a --sufix allows us to create a custom version which
clearly reflects the fact that this version is the result of some
special situation. In our example, I might specify a --suffix r2
or something similar, indicating that this version diverged from
the main trunk at some point.